Implemented stamina system
I have created a new script called "StaminaComponent" which you can attach to any gameobject you want to have stamina. This script handles the consumption of stamina from the gameobject, stamina regen, and the function that you will call from the main character script returns an enum for the strength level of the ability you're using. Simply call the "ConsumeStamina" function in a switch statement, and implement the functionality for each strength level.