Player revive mostly working
Swapping from our initial approach for player reviving where we would spawn a Corpse prefab of the specified character that had all the inbuilt functionality, I have now changed to what I think is a better approach (for now) where the player is built from a ragdoll skeleton but with the bones being kinematic, they are not affected by forces and can be instead driven by animation. When it comes time to die and spawn a body, I simply spawn a new Player prefab and prepare the current player object to become a corpse by destroying all the player functionality components, enabling/adding corpse functionality components and setting all bones to be non kinematic. I feel this is better because we wont have to create 2 prefabs for each character type and make sure we spawn the right one.