Fixed input bug
Fixed bug where player could no longer use the Unity Input System after reviving. This was due to no devices being available for the new player to use as the old Player Input component was still using it. To fix this I simply first copied the old player input asset so I could assign it to the newly spawned player, then I set the old Player Input's Input asset to null so a device would be free for the newly spawned player, finally I set the newly spawned player's Player Input components' Input asset to the old one that I had stored. One issue is the corpse ragdoll is not following the parent object. Another thing to remember is to set the corpse parent's joint projection variables and the break force.