Fixed issues with holding corpse after revive
After making sure to add the correct components and initialising the right variables, I came across a problem of the body being sent flying and being held in the wrong position. The problem was that the mass of the PickupPosition that the body would be joint to, was way too big because after trying to set it to 0, it is instead set to "1e-07" which is apparently equal to a very big number. I simply changed the mass to 1.