Fixed a few revive bugs
The players HUD has a reference to the player gameobject in the scene so that it can track the players relevant stats such as health and stamina but after spawning a new player, this reference would still be the old one. I simply made a function that replaced the old reference with the new player character.