Added block without held corpse, and added/fixed interactable objects
The player is now able to block with the left mouse button (I also changed throw to middle mouse button to keep the main abilities easily accessible).  If the player blocks with enough stamina above 0 they do a full block: consuming stamina but dealing no damage, if the player blocks with stamina but it goes below 0, the player does a half block and takes half damage, if the player does not have enough stamina to block, they still can, they take half damage but their stamina regen delay is doubled. Fixed door interactable and minor tweaks to ragdoll values that messed with interactions. Added spiked object/damaging area that can be placed in the environment.