Added basic damaged and reviving visuals and parry
The IDamageable interface now has a DamageFlash function which, when implemented causes the damaged object to flash a different colour (white). When the player is in the reviving state their material is changed to signify that they are in the reviving state and have limited options. Also added a parry mechanic that is active for a duration at the start of blocking which instead of damaging the players stamina or health, gives them stamina back for a successful parry. A problem I can see with the current implementation is that the player can just spam the parry/block button when not attacking to basically never take damage. Will have to think of a fix for that.