Returned to rigidbody controller and fixed moving platform
Due to issues with how the character controller interacts with objects and values in the game, I have switched back to using a rigidbody for the player controller. I have also fixed a couple issues regarding the moving platform that had been plaguing me. These included objects not being able to be made children of the platform (caused by the CharacterController) and, when spawning a new player for the revive, their new gameobject would be sent to a different location and their skeleton messed up (this was due to scaling of the platform). To fix this second issue, I simply made the mesh (which will be scaled) a child of the moving platform prefab which I would be making objects children of. Because this parent is not being scaled, it wont affect the new children negatively.