Player can now damage enemies with thrown corpses
The player now has the ability to deal damage when  the corpse they throw hits an enemy. Current issue with throwing is that the thrown body does not always properly go in the direction you want so I will have to figure out the best way to make the throw feel nice to use.