Added damage knockback, fixed throw and ai bugs
Before, the held corpse would be thrown from their current position (being held by the player) however I found that it caused the body to not go completely in the intended direction, making the throw hard to use. Instead, now the body is teleported directly in front of the player as force is applied so the body will be thrown in the exact direction the player is aiming. For the AI, originally there were no issues however after when attempting to add damage knockback I found that it conflicted with the NavAgent components movement system. I made some changes to only use the NavAgent to find the path the enemy must take, and I am now using the rigidbody to move the enemy along the path. There were a few other AI related issues revolving around the enemies attack stopping distance but after some code clean-up and refactoring, it should work as intended.