Created working combo attack system
Created various assets to contribute towards a combo attack system that works successfully for the player (needs implementing for enemies). This system involves an "AttackComponent" script that makes use of a list of "Attacks" containing various pieces of data such as the attack animations name, speed, and damage which are initialised through scriptable objects. In the editor, it is possible to create new scriptable objects for different attacks and add them to the combo list to give the parent of the AttackComponent, combo capabilities. This system will also be able to be used with single attacks as you can simply keep the combo at 1 attack.