Both player and enemies now have properly implemented "combos"
Both the player and enemies can now be given "combos" which can consist of 1 or more attacks. The logic for attacks is mostly driven through animation events and the AttackComponents "StartAttack" function simply starts the appropriate attack animation. I have had a big focus on making things more data-driven and modular so that similar things are in the same place (not having to copy similar code around), and to change values, you won't need to run around the file structure looking for the correct script.