Fixed gravity issues and ai pathfinding visuals
Due to the implementation of movement for the player and enemies, when falling and moving, gravity was not being applied due to y velocity being set to 0 when moving. This was fixed by constantly applying the gravitational force velocity on the y axis and using the current rigidboy y velocity (the one with gravity applied) when figuring out the final movement velocity. There was also an issue with the enemy ai stuttering when moving. To fix this, I simply made the enemy ai follow the current path at smaller intervals to make it smoother.